refactor: move yours-sincerely-specific wrappers out of @repo/ui
Drop app-specific helpers from packages/ui back to apps/web:
- ProfileAvatar → apps/web/src/components/profile-avatar.tsx
- Form/FormField/FormItem/FormLabel/FormControl/FormMessage → apps/web/src/components/form.tsx
- dropdownMenuItemVariants → apps/web/src/lib/drawer-item.ts (renamed drawerItemVariants)

Also:
- command: move DialogHeader inside DialogContent, strip built-in check icon (consumers can add)
- sonner: drop custom theme-name mapping (light-purple/dark-purple map via Sonner's system theme)
- dropdown-menu: inline the item className (no cva export)
- avatar: drop ProfileAvatar from shared components
- globals.css: refresh tokens

Add @radix-ui/react-slot, @hookform/resolvers, react-hook-form to apps/web now
that form.tsx lives there.
