+++ apps/web/package.json
@@ -17,7 +17,9 @@
     "@capacitor/splash-screen": "^8.0.1",
     "@knocklabs/react": "^0.11.11",
     "@lottiefiles/react-lottie-player": "^3.6.0",
+    "@hookform/resolvers": "^5.2.2",
     "@number-flow/react": "^0.6.0",
+    "@radix-ui/react-slot": "^1.2.4",
     "@react-hook/hotkey": "^3.1.0",
     "@repo/api": "workspace:^",
     "@repo/db": "workspace:^",
@@ -32,6 +34,7 @@
     "next": "catalog:",
     "next-themes": "^0.4.6",
     "react": "catalog:",
+    "react-hook-form": "^7.72.0",
     "react-dom": "catalog:",
     "react-infinite-scroll-hook": "^6.0.1",
     "react-use": "^17.6.0",
