+++ apps/web/src/app/(app)/posts/_components/more-button.tsx
@@ -18,7 +18,7 @@ import {
   DrawerTitle,
   DrawerTrigger,
 } from "@repo/ui/components/drawer";
-import { dropdownMenuItemVariants } from "@repo/ui/components/dropdown-menu";
+import { drawerItemVariants } from "@/lib/drawer-item";
 import { toast } from "@repo/ui/components/sonner";
 import { useMediaQuery } from "@repo/ui/lib/utils";
 import { useMutation, useSuspenseQuery } from "@tanstack/react-query";
@@ -92,7 +92,7 @@ export const MoreButton = ({ post }: Props) => {
 
   const buttonClassName = isDesktop
     ? "rounded-sm p-8"
-    : dropdownMenuItemVariants({ className: "w-full justify-start h-10" });
+    : drawerItemVariants({ className: "w-full justify-start h-10" });
 
   const menuItems = [
     <Button
