+++ apps/web/src/app/(app)/profile/_components/profile-form.tsx
@@ -2,8 +2,8 @@
 
 import { zodResolver } from "@hookform/resolvers/zod";
 import { updateUserInput } from "@repo/api/user/user-schema";
-import { ProfileAvatar } from "@repo/ui/components/avatar";
-import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from "@repo/ui/components/form";
+import { ProfileAvatar } from "@/components/profile-avatar";
+import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from "@/components/form";
 import { Input } from "@repo/ui/components/input";
 import { toast } from "@repo/ui/components/sonner";
 import { useMutation, useSuspenseQuery } from "@tanstack/react-query";
