+++ packages/ui/src/components/avatar.tsx
@@ -90,30 +90,4 @@ function AvatarGroupCount({ className, ...props }: React.ComponentProps<"div">) 
   );
 }
 
-type ProfileAvatarProps = {
-  className?: string;
-  displayName?: string | null;
-  src?: string | null;
-  alt?: string;
-};
-
-function ProfileAvatar({ className, displayName, src, alt = "" }: ProfileAvatarProps) {
-  const initials = displayName?.slice(0, 1) || "?";
-
-  return (
-    <Avatar className={cn("size-9", className)}>
-      <AvatarImage className="dark-purple:invert dark:invert" src={src ?? undefined} alt={alt} />
-      <AvatarFallback className="animate-in fade-in uppercase">{initials}</AvatarFallback>
-    </Avatar>
-  );
-}
-
-export {
-  Avatar,
-  AvatarImage,
-  AvatarFallback,
-  AvatarGroup,
-  AvatarGroupCount,
-  AvatarBadge,
-  ProfileAvatar,
-};
+export { Avatar, AvatarImage, AvatarFallback, AvatarGroup, AvatarGroupCount, AvatarBadge };
