+++ packages/ui/src/components/sonner.tsx
@@ -10,18 +10,12 @@ import {
   Loader2Icon,
 } from "lucide-react";
 
-const sonnerTheme = (theme: string | undefined): ToasterProps["theme"] => {
-  if (theme?.includes("dark")) return "dark";
-  if (theme?.includes("light")) return "light";
-  return "system";
-};
-
 const Toaster = ({ ...props }: ToasterProps) => {
-  const { resolvedTheme } = useTheme();
+  const { theme = "system" } = useTheme();
 
   return (
     <Sonner
-      theme={sonnerTheme(resolvedTheme)}
+      theme={theme as ToasterProps["theme"]}
       className="toaster group"
       icons={{
         success: <CircleCheckIcon className="size-4" />,
